libtracker-miner: don't crash when first indexing files out of /home
If tracker_miner_manager_index_file() is called with a path outside of
the home directory, we'll end up recursing up to the file system root,
to create the parent directory objects.

When the tree we're passed is the filesystem root, and the file is
exactly equal to it, we should return the tree root immediately, or
we'll crash later in the while loop.

https://bugzilla.gnome.org/show_bug.cgi?id=694783