all: Improve start/stop logging for all processes
Now we have Starting/Stopping in the log_init/shutdown functions not just the
process and version.

We also only print the log filename if we're writing to a log file at all.

We now don't print to stdout that we're initializing either.

The TrackerConfigFile g_message() line about which file we're "monitoring" is
removed too, this was always the first line and quite redundant anyway now we
use gsettings.