Fix AS ?foo handling in FTS queries
FTS queries implicitly add an "AS var" clause to the translated
SQL select query so values can be matched with the outer query that
accesses FTS tables, which resulted in doubly added AS clauses if
it was specified explicitly in SPARQL too.

So, make sure the clause is just added once.