tracker-extract-pdf: fix uninitialized value
'bytes_remaining' needs to be initialized to the 'bytes_expected' value, so that
the proper amount of remaining bytes is kept during the outer while() loop. In
the first iteration, this will be -1, which will then be updated to afterwards
to the same 'bytes_expected' value read from the stream.

https://bugzilla.gnome.org/show_bug.cgi?id=702221