Revert "tracker-extract-mp3: Do not always use guessed encoding for ID3v2"
The new ICU based encoding detector should be more reliable, plus this hack
is triggered consistently on a very wide set of locale environments, also
happening for legit files that do need the encoding. The bug and file that
motivated this patch are also long lost and forgotten... so this patch will
remain in git history.

This reverts commit 9b2acf60b54fe9332a48f5260350c050e65c4111.