test,libtracker-fts: allow alternate number of expected words
In ICU 50 the word splitting algorithm changed, so this test may get
different results depending on the host environment. The test simply
needs to verifying that word splitting works, so it should allow either
value.

As of November 2013 Ubuntu is still shipping an older version of ICU so
it would be awkward to simply increase the version we depend on.

https://bugzilla.gnome.org/show_bug.cgi?id=699412