Avoid starting tracker-store twice
We install both of:

/etc/xdg/autostart/tracker-store.desktop
/usr/share/dbus-1/services/org.freedesktop.Tracker1

If something else speaks to tracker via during login, what will happen
is that dbus-daemon will spawn one, and gnome-session will also spawn
one, and one of these will emit a critical message.

To fix this, change the .desktop file merely use DBus (it'd be nice if
there was an easier way to do this).

Instead, maybe we could consider only dbus-activating tracker?  In
other words, if nothing wants to talk to us, maybe we shouldn't be
starting by default...

https://bugzilla.gnome.org/show_bug.cgi?id=715040