libtracker-miner: Add TrackerDecorator
This object is a TrackerMiner abstract implementation that instead of
actively looking for information, it passively receives notifications
through the tracker-store GraphUpdated DBus signal, and offers the
possibility to fill in further information of newly added resources
or older resources missing inspection by this miner.

By default, this object receives information from all resources with
tracker:notify true, the class-names property can tweak this so it
only acts on events from a certain rdf:type.

In order to know which elements are missing inspection, the miner
relies on a nie:DataSource specific to each TrackerDecorator
implementation. On startup, TrackerDecorator will check for the
eligible elements missing that datasource to ensure those are checked.
The data source can be set through the data-source property.