tracker-control: Added option --collect-debug-info
Useful when debugging problems to diagnose the state of Tracker on
your system. The data is output to stdout. Useful if bugs are filed
against the project itself.

Data collected includes Tracker version in use, disk space available,
size of the databases on the disk, the configuration in use, states of
the index (e.g. last filesystem crawl, data set locale, etc.) and
finally statistics about the data in the database (e.g. how many
nfo:FileDataObject resources exist).