extract: Plug a leak
If the string being processed is detected as UTF16, the "utf8"
variable would already contain a malloc()'ed string, so the
GString would be leaked altogether.

Then, ensure the string is freed if already converted from UTF16,
and spare the utf8 validation too, since we already know the string
is valid utf8.