libtracker-miner: Turn thumbnailer into an object
TrackerMinerFS in libtracker-miner was already doing most of the
tracker_thumbnailer_* calls necessary. The only API required on
callers only was tracker_thumbnailer_init/shutdown(). So just
turn this into an object so all usage is kept private to
libtracker-miner, and we can remove these headers out of the
public.