libtracker-miner: Add TrackerMinerOnline
This TrackerMiner is a simpler replacement for TrackerMinerWeb that
doesn't get into credentials handling. It handles network state,
emitting ::connected or ::disconnected on the way, and ensuring the
miner is paused/resumed as necessary when suitable networks come and
go, implementations of this miner can control this behavior through
the return value in the ::connected signal.