TrackerDecorator: Sort elem_queue to extract important files first
There are 2 things affecting an element position in the queue:
prepend (typically files from an USB key the user plugged),
and prior (the file type matches what applications needs first, e.g.
nfo:Audio).

The queue is sorted like that:
1) prepend and prior files
2) prior files
3) prepend files
4) the rest

This garantees that if I open the Music app and plug an USB key, I'll
get the MP3s from that key first, which is probably want the user
wants.

https://bugzilla.gnome.org/show_bug.cgi?id=719802

Conflicts:
	src/libtracker-miner/tracker-decorator.c