functional-tests: Don't rely on just tracker-control -t to kill tests in time
The command `tracker-control -t` simply kill(2) all tracker processes and
exits. It does not wait/check the processes to terminate. In some test cases
it happens that the test case tries to launch its own instance of some tracker
process, but it fails because the old process is still there.