functional-tests: Don't require miner-fs being idle for testing data inserted
In many test cases the tracker_miner_fs_wait_for_idle() call is used to detect
miner operation has completed. It returns when miner's status changes to
"Idle" (or on timeout). Unfortunately at the time miner goes idle it is not
guaranteed the data are already in store - and the related test fails.

The test case 301-miner-resource-removal.py does it a better way - it listens
to GraphUpdated signal sent by store and waits until the desired resource is
announced being added or removed.

There is comment inside 301-miner-resource-removal.py: "FIXME: put this stuff
in StoreHelper". This patch is to follow that comment.

Needed by tracker-tests-310-fts-indexing-use-graph-updated-signal.patch.