functional-tests: Use GraphUpdated signal from store, not miner-fs Idle status
The tracker_miner_fs_wait_for_idle() call is used to detect miner operation
has completed. It returns when miner's status changes to "Idle" (or on
timeout). Unfortunately at the time miner goes idle it is not guaranteed the
data are already in store - and the related test fails.

A better way is to listen to GraphUpdated signal sent by store and wait until
the desired resource is announced beng added or removed.

Depends on tracker-tests-allow-reuse-graph-updated-signal-handling.patch.