tracker-extract: Remove the need for fork() and sub-processes with PDF extraction
This fuctionality we're removing has been around since
be58d8da6da5a8f16ca83f78a9427ca4de723151 where Philip added code to kill a
forked process after n seconds for really complex PDFs (or PDFs that take
longer to process on embedded devices). This was covered under NB#290406.

Later I improved the code to use select() instead of signal handlers due to
IO blocking in some situations in commit
9c4e166ec101c79bdeac30696371ffe0abd4e046 and as part of GB#680897.

However, the extra complexity is no longer needed as far as I can tell. There
was a good reason for the IO blocking in the past and I believe it has been
fixed by Carlos in one commit or another and with the new GTask and
_run_in_thread() APIs we have now, it's unnecessary to use this approach. IF
anything, we should add some timeout or cancellation in the task issuing
(main?) thread instead of implementing this for each extractor. Currently we
don't do this though.

As an additional note, this may fix GB#726421.

I've tested this patch with files attached to previous bug reports related,
including GB#680897 and GB#685378.