Revert "libtracker-sparql: Fix undefined reference to symbol 'tracker_ipc_bus'"
This reverts commit 4623bcd6b089be5c32d615a1733bbcd387a7420e.

It's not libtracker-sparql that requires this symbol to be defined, it's
binaries using the symbols from libtracker-common that should have this fix
instead. Use of the TRACKER_IPC_BUS and above function is internal only.

This error was highlighted with --as-needed being used with the linker.

This is properly fixed in commit: 963e53eb79a5b727ef008a875f16005e4c6d67dd