tracker-extract: MP3s with ID3v23 tags should not hit infinite loop
The extraction was using a while() loop with pos < size, but would never be
FALSE and continued forever. The size is the size of the file, but actually,
it should be using tsize, which is the tag size.

What was happening, was the while() loop would increment over sizes of +10
+content-size, but if there was padding, there was no tag frame and so no
content size and the increment of 10 eventually would step into the RAW MP3
data.

Fixes:
https://bugzilla.gnome.org/show_bug.cgi?id=728252

Specification details:
http://id3.org/id3v2.3.0