libtracker-miner: Create TrackerFileSystem in constructed() not init()
We do this because the GFile which is the root for the TrackerFileSystem is a
property which is set on object construction. If we try to do this in init()
the root is unset at that point. So we wait until constructed where properties
are guaranteed to be set.

Also Chain parent constructed() in TrackerFileNotifier