libtracker-miner: Create TrackerIndexingTree in constructed() not init()
We do this because the GFile which is the root for the TrackerMinerFS is a
property which is set on object construction. If we try to do this in init()
the root is unset at that point. So we wait until constructed where properties
are guaranteed to be set. We need this to give to the TrackerIndexingTree that
requires a root argument if the file system is not 'file:///' based.