functional-tests: Fix tests that use ExtractorHelper
The DBus name of the tracker-extract process has changed, so some of the
functional tests were hanging forever as they were watching the wrong bus name.

There's also no longer a tracker extract interface at all.

This fixes a bunch of tests that were hanging forever waiting for the
extractor's object to appear.