functional-tests: Rework 400-extractor.py test
Tests now call the tracker-extract process manually and parse its
ouput. This is necessary because the D-Bus interface it used to expose
has been removed. It'd be better to run the tests with a store and
miner and check that the correct data is inserted in the store, but
for now this approach is OK, and it does have the advantage that the
extractor is tested independently of the store & filesystem miner.