functional-tests: Rewrite DConf code to use GObject introspection
The tests will now detect if the DConf backend is not in use and raise
an error, where previously tests would fail for confusing reasons. Also,
we no longer have to shell out to 'gsettings' to set up the
configuration.

Configuration values now have to be specified as GLib.Variant instances
instead of Python types. That's a bit of a pain. PyGI should have a way
of creating a GVariant from an arbitrary Python value, but I didn't find
one.