functional-tests: Replace use of PyGObject and PyGTK with PyGI
If we do 'import gobject' anywhere then we cannot use the new
GObject-introspection based Python bindings, because they conflict.

This change means the functional tests can now use PyGI, instead of the
deprecated and out-of-date PyGObject and PyGTK.

Note that the dbus-python bindings are still widely used in the
functional-tests. These work fine but are deprecated. We should
switch to using GDBus through GObject-introspection instead when possible.