functional-tests: Fix and improve 601-applications-sync
This test adds data that would normally be added by the FS miner.
In order for the resource to be correctly deleted when the file is
removed, it must set nie:isStoredAs correctly.

Replaced a tracker_miner_fs_wait_for_idle() call with await_resource_deleted()
too, and added a note about the error tracker-miner-fs logs during this test.