functional-tests: Allow waiting on a specific property, and fix tests that need this
Resources are created when the miner process sees them, but often we
want to wait until they have been processed by the extractor, too. We
can achieve that by waiting for the insertion of a specific triple that
we know is set by tracker-extract.

This required adding a new parameter to the await_resource_inserted()
function.

The 310-fts-indexing and the base writeback test have been fixed to make
use of this, instead of using tracker_miner_fs_wait_for_idle().