Always guarantee metadata, remove --enable-guarantee-metadata option
The --enable-guarantee-metadata flag ensures that nie:title and
nie:contentCreated will always be set for a given file, even if they
need to be guessed based on the filename and mtime.

It was previously disabled by default, although the functional tests
rely on it being enabled.

There should be no harm in making this behaviour the default, and
removing the option. If this turns out to have unintended consequences,
we should either fix the code in question, or remove it completely.

https://bugzilla.gnome.org/show_bug.cgi?id=733536