tracker-extract: Implement "failsafe" extraction
The extractor will store a symbolic link in a user directory in /tmp to
the file(s) being currently processed, encoding also the number of tries
previously performed. If the extractor happens to crash, the links will
be used on a future respawn to resume operation.

If enough crashes happen on a given file, the file will be then discarded,
and tagged with an "extractor-failure-data-source" nie:dataSource,
additionally to the dataSource that will make tracker-extract take the
file as indexed in future runs.