tracker-writeback: Remove broken code to set file attributes
Remove some code which tries to set attributes of a file which no
longer exists.

This code was added in the following commit:

    commit 16fe51de1e93016a122ab089b0e0201c18de7ebc
    Author: Philip Van Hoof <philip@codeminded.be>
    Date:   Wed Nov 9 12:56:59 2011 +0100

        tracker-writeback: Preserve file permissions

        Fixes NB#289953.

Seems that the intent was to preserve the permission bits of the
file after tracker-writeback rewrites it. The code already takes
care of this at the time the temporary file is created.