tracker-extract: Use default signal handler for SIGALRM and SIGABRT
The ALRM handler was introduced in commit
d9d1881c2548e5d6d55fad1e897f8f058ef28696. The alarm() function seems to
no longer be used, so we can remove it.

The ABRT handler was introduced in commit
3f42a4390d48c6a4840a7bcfce74673ebfd23479. I'm not sure of the reason.
Other Tracker processes use the default handler for ABRT, so let's be
consistent.

https://bugzilla.gnome.org/show_bug.cgi?id=734144