Processes shouldn't install the signal handler until the main loop starts
The signal handler we install calls g_main_loop_quit() when SIGTERM or SIGINT
are raised, but this is only useful if the loop was started. Sending SIGTERM or
SIGINT to a Tracker process while it is starting up fails to stop the process,
and triggers the following a warning:

  GLib-CRITICAL **: g_main_loop_quit: assertion 'loop != NULL' failed

Ideally we'd call initialize_signal_handler from an idle once the main
loop has started, but installing it just before the loop starts should
be a big improvement.

https://bugzilla.gnome.org/show_bug.cgi?id=734144