functional-tests: Remove all remaining 'wait for miner to be idle' calls
We are running against a clean data directory when run by 'test-runner.sh', so
the miner should not have any initial crawling work to do. Therefore the tests
don't need to wait for it to be idle before starting.

Test cases should use the await_resource_inserted() and
await_resource_deleted() methods as these are (in theory) reliable. Waiting for
the miner to go idle and hoping this means that the desired file was crawled
and extracted is not reliable at all.

This should hopefully make the tests robust enough to be run continuously
without random failures occuring.

The tracker_miner_fs_wait_for_idle() functions are gone forever!