tracker-stats: Change default behaviour and add new command line options
Now we only show common RDF types by default, e.g. nfo:Document and it is
based on .rules file fallback types and some others specifically added.

There is also now a --all option to return all stats as was the behaviour
before.

In addition to this, the EXPRESSIONS provided on the command line following
the OPTIONS can be used to filter the stats, e.g. : 'tracker-stats doC' will
list the nfo:Document class and resources in the DB matching.