data: Merge all miner .desktop files into existing DBus .service files
So .service files store just the 'Name' and 'Exec' information for the DBus name
and binary to instantiate. Now we keep the 'Path', '_DisplayName' and
'Comment' in here too. These all come from the data/miners/*.desktop files
which no longer exist and are mainly used for tracker-control to know about
miners that are not running and that exist or have been installed.

The old '_Description' is now '_Comment'.
The old '_Name' is now '_DisplayName'
The old 'DBusPath' is now 'Path'.

As usual all '_' prefixed keys are translated so tracker-control still shows
translated miners in its output.