gvdb: Don't use TRACKER_UTILS_{CFLAGS|LIBS}
Now these no longer exist, we can't use them and we shouldn't have been
anyway. This updates the test to not use GIO APIs anyway, it was only needed
to remove a file and it's not needed by the GVDB library so we don't want to
add the dep just to fix a unit test. Instead, we now use the unistd unlink()
API.