libtracker-miner: _NO_STAT breaks in FileNotifier without a GFileInfo
GFileInfo is ABSOLUTELY required in the Crawler, without it the
TrackerFileNotifier will that top level roots have been deleted because the
GFileInfo GQuark does not exist.

This is seen easily by mounting a removable device, indexing, then removing,
then re-inserting that same device.

The check is done later in the TrackerFileNotifier by looking up the qdata
that we set in both conditions below.