all: Migrate to GKeyfileSettingsBackend and generate man pages from schemas
Now we've removed the internal key file object, we've had to put in place
another method for supporting the existing TRACKER_USE_CONFIG_FILES
environment variable. Thanks to the GKeyfileSettingsBackend provided by
GLib, we can fallback to old school INI type config files for embedded
solutions or cases where we don't want dconf as a backend. This works rather
well.

IT should be noted, the INI files are *NOT* written out in full if they do not
exist, only options which are saved or different to the default settings are.
This is how it should be too.

Now we build man pages based on GSettings schemas using xsltproc with the
template in docs/manpages/gsettings.xsl. This is a useful aid when trying to
understand what config files can have in them. One thing it does highlight, is
the config documentation could be better :)