tracker-miner-fs: Delete nie:dataSources on update
Decorators should be updating again all relevant information, so ensure
their nie:dataSource is deleted for these to kick in again.

This used to be unnecessary because updated items just went to the
internal decorator queue, without further checks. Now that the decorators
query the database for the next items missing that dataSource, it is
crucial to do this somewhere.