TrackerFileNotifier should only care about nfo:FileDataObject resources
The motivation for this commit is that the 600-applications-camera.py
functional-test would trigger the following warning:

   (tracker-miner-fs:28346): Tracker-CRITICAL **: tracker_string_to_date: assertion 'date_string' failed

The cause was in tracker-file-notifier.c:sparql_files_query_cb(), which
processes a query that expects every resource with an nie:url property
to also have nfo:fileLastModified property.

Not all resources correspond to files. The TrackerMinerFS should only be
processing resources that have type nfo:FileDataObject.