Fix max_depth parameter causing contents of a directory to be ignored
Since commit 41e6ab84e7a540b20f8173c2bbaee11de6ef381f, the contents of
directories listed in 'index-single-directories' is be ignored. The
directory itself appears in the store but the contents do not.

The intention of passing max_depth=1 if TRACKER_DIRECTORY_FLAG_RECURSE
is false must be to prevent unnecessarily processing any subdirectory
nodes that are hanging around in the TrackerFileSystem. However, the
documentation of g_node_traverse says:

    "If depth is 1, only the root is visited. If depth is 2, the root
    and its children are visited. And so on."

Thus, we must pass max_depth+1 to the tracker_file_system_traverse()
function so that the files in the directory are processed as well as
the directory itself.

See also: 243c97fa96a2f6b19bce1acf26946d1a516a0a1c which introduced the
'max_depth' parameter as a way of reducing the number of GFile objects
in memory at the same time during crawling.