functional-tests: Improve stopping of child processes
Instead of using `tracker daemon --kill=miners` to stop the miners, use
the Python 'subprocess' module. This ensures only the correct miner
process gets killed if more than one is running. (I also found that
`tracker daemon --kill-miners` was hanging my system, which was the real
motivation for fixing this, but I have no idea why that happened).

If a process fails to terminate after we sent it TERM, it now gets sent
KILL. Previously the test would wait forever.