libtracker-miner: Restrict the amount of data that is logged for errors
SPARQL update strings can be very long if the entire text of a document
is being stored. Dumping these huge strings to the log eats up disk
space and makes them harder to follow.

However, often, the occurence of such an error indicates a broken
extractor. In those cases, knowing part of the SPARQL can help in
identifying the file that triggered the error.

Usually the "nie:plainTextContent" property is the last one in the
string, so we truncate the error messages at the first occurance of
this property to achieve the best of both worlds.

https://bugzilla.gnome.org/show_bug.cgi?id=735406