tracker-data-manager: Re-creating tables don't add any property column
If, during database update, the migration code chose to create a table from
scratch, dumping data from the original one (eg. removing columns for
properties that now have maxCardinality>1 and move to their own table),
the destination table would be missing virtually all fields.

When this happens, in_update differs from in_alter (the former tells whether
we're creating the DB from scratch, the latter whether we're altering the
original table). As we're actually creating the table from scratch, we must
behave as such when dealing with the affected properties.

https://bugzilla.gnome.org/show_bug.cgi?id=745737