file-notifier: Exclude pending dirs from the contents sparql query
Directories added to the pending queue (ie. at the edge of max-depth)
may also be signaled as "updated" and added to the updated files array.
As these directories will be crawled/queried independently later on,
it doesn't make sense to include these in the current query for
directory contents.

This fixes spurious reindexes seen across tracker-miner-fs restarts,
an(y) updated folder would be added on both queue/array, the sparql
query for directory contents would include that folder that hasn't
been crawled yet (hence being in the pending queue) and incorrectly
signal all its contents as deleted. The folder would be just reindexed
again from scratch on the next restart.

https://bugzilla.gnome.org/show_bug.cgi?id=741852