libtracker-miner: Keep a monitor on root index tree files on deletion
If Tracker is running with some index-recursive-directories set (for
example, ~/Music), and one of those directories is moved, then the file
monitor watching for its existence is deleted. This means that if it is
then moved back to the location listed in index-recursive-directories,
its renewed existence is not detected, and its contents are not
re-indexed.

Fix this by keeping a monitor around on a directory if that directory is
listed as an index tree root, even if the directory is deleted.

https://bugzilla.gnome.org/show_bug.cgi?id=750394