tracker-miner-fs: Fix a completion check when removing the final task
Depending on how mining goes, this path might be the last one taken
before the miner is ready to go idle again. However, the check on the
task pool size is guaranteed to be false because the task which
item_add_or_update_continue() was called on has not yet been removed
from the pool — that’s done directly below.

Fix that by removing the task from the task pool before checking whether
the pool is empty.

This fixes stalls in tracker-miner-fs where `tracker-control -S` would
show the following for ever (when running with
index-recursive-directories set to a non-empty list):
   1%  File System - Crawl finished for directory 'blah'

Previously, the only way to fix this was to pause and then resume the
miner.

https://bugzilla.gnome.org/show_bug.cgi?id=751172