extract-mp3: Account for the tag header size in calculations
In the parsing of id3v2[0-4] tags, we calculate offsets in base
of the size we get in the header, which according to id3 specs,
doesn't include the 10 bytes of the header itself.

This means there's a disparity of 10 bytes between our 0-based
offsets, and our 10-based fetched size, we must account for this
when iterating through frames.

Also, we were possibly (and mistakenly) iterating till the end
of the file (comparing against size instead of tsize(+header) in
the while loops), we only have to parse one id3 block at a time.
If the former bug could conceivable make us miss the last frame,
this one would overcompensate for it.

Less importantly, this fixes the incorrect "position +
content size > tag size, not processing any more frames" debug
spew I was seeing on many mp3s with -v 3.