extract-mp3: Stop processing after finding padding
ID3v2[0-4] Allows the possibility of some padding at the end of an
ID3 block, defined to be filled with 0x00, just in case later tag
additions don't cause a full rewrite of the file. This means there's
nothing for us when we hit this.

ID3v2.4 also defines the existence of "ID3 footers", those always
start as "3ID", and should be skipped for our purposes, as those
are basically a mirror of the ID3 block.

This most notably avoids a bunch of "Ignoring unknown frame ''"
spew on -v 3 logs, as padding is fairly common.