extract-mp3: Bail out on encoding detection if confidence is too low
Libicu encoding detection is able to tell the confidence it got on
the detection, we should be using that in case the confidence is
too low, as that means the returned encoding is probably bogus, and
we have an encoding to fallback on.

This fixes detection on the file reported on bug #735515, where
a couple of 'ï' chars (valid ISO-8859-1) make libicu detect UTF-16BE,
although with an extremely low confidence.

https://bugzilla.gnome.org/show_bug.cgi?id=735515