tracker-extract-gstreamer: Fallback to preview image for album art
Some files don't provide a cover image, but just a preview image, we can
use that as albumart.

This commit reuses code from a fallback which looks bogus nowadays, since
it just does the same than the while loop above. If we reached there, it
would fail again for sure.

https://bugzilla.gnome.org/show_bug.cgi?id=732236