libtracker-miner: Fallback to basename checks on hidden files
When we receive DELETED events on files, the TrackerFileNotifier/
TrackerIndexingTree attempt to find out if its hidden and should
thus be ignored. However the g_file_query_info() evidently fails
at that time, with the file already deleted, so we think the file
is not hidden and the delete operation gets propagated to the
TrackerMineFS.

Fix this by resorting to basename checks as a last resort, so
we do the right thing even after the file is gone.