Revert "libtracker-data: delete elements from the Resource table"
This reverts commit ae3ab553d0e5faf0dadc85586dac0c0ee4443394.

This approach is a bit too drastic, as it may confuse queries triggered
by the deletion of the resource, the SQL translation of DELETE queries
using the deleted resource in the WHERE will fail because there is no
matching URI for that row ID anymore.

Another less drastic approach is done in the next commit.