rss: Perform extraction/insertion of feed items at once
The async ASK/INSERT pattern per-feed was vulnerable to nie:url collisions
if we happened to process several feed items pointing to the same URL
simultaneously.

Change this so we first query the older items, and then we check items are
really up to date, inserting only the newest of occurrences, and updating
the database if the content changed for real (eg. the post was updated).
This is done in a single update_array operation for all feed items in a
channel.