libtracker-extract: Add builtin dummy extractor
For rules with no ModulePath key (i.e. those we only want to set
rdf:types for) we relied on the lack of generic extractors that
might handle the mimetype.

If there was one such generic extractor, the TrackerMimetypeInfo
iterator would skip the empty rule and directly attempt the generic
extractor.

Because it is conceivable that no-op is desired for some files,
we shouldn't attempt to skip it, nor run further modules after the
dummy rules, so just give this dummy fastpath a presence that
tracker-extract will make sense of.