tracker-extract: Use dummy fastpath for svg extraction
The gstreamer based extraction is less than useful for our purposes,
it could spend considerable amounts of cpu time and memory in
building a pixel representation of the image (which is useless for
us), and would extract nothing relevant from the file (w/h are
meaningless on scalable images, other details such as geolocation
just don't apply here).

So SVG through gstreamer was a very expensive no-op, it can be
safely made to use dummy extraction, and it will still be set
the corresponding rdf:types.

This is our side of the fix for:
https://bugzilla.gnome.org/show_bug.cgi?id=748608