functional-tests: Improve 'manual start' mode
This mode now allows you to run the tests against already-running
Tracker processes. A lot of tests depend on specific miner-fs
configuration so they might not work, but it can be useful.

You can still start the test and then manually start the Tracker process
if you want.

Note that enabling manual start on the commandline probably doesn't
work (commandline options for these tests are pretty broken) so just
change the options.is_manual_start() function to 'return True' if you
want to use this mode.